Back to npot textures
Since there was no difference between using pot or npot, we are back to
npot since this reduces CPU->GPU bandwidth usage, which is one of the
bottlenecks for nibble at the moment.

Also increased the use of integer arithmetric in both the HLSL and the
GLSL shaders, to avoid floating point errors, specially in older
hardware.

Will check if those changes fixed the issue with the with background
when running in old intel graphics hardware (which we still don't know
the exact root cause).